feat: define remote fetch function to RestaurantAPIFetchable protocol
